# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.envelope:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  handler.errorEnvelope:
    properties:
      errors:
        items:
          $ref: '#/definitions/handler.fieldError'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  handler.fieldError:
    properties:
      detail:
        type: string
      field:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://github.com/404nfidv2/go-nuxt-starter-kit
  description: REST API for the Go-Nuxt Starter Kit (auth, profile, notifications,
    admin).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go-Nuxt Starter Kit API
  version: "1.0"
paths:
...
  /auth/2fa/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: TOTP code
        in: body
        name: body
        required: true
        schema:
          properties:
            code:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.envelope'
            - properties:
                data:
                  properties:
                    recovery_codes:
                      items:
                        type: string
                      type: array
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Confirm 2FA setup with TOTP code — returns recovery codes
      tags:
      - 2fa
  /auth/2fa/disable:
    post:
      consumes:
      - application/json
      parameters:
      - description: Current password and TOTP code
        in: body
        name: body
        required: true
        schema:
          properties:
            code:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Disable 2FA
      tags:
      - 2fa
  /auth/2fa/recovery-codes/regenerate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Current password and TOTP code
        in: body
        name: body
        required: true
        schema:
          properties:
            code:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.envelope'
            - properties:
                data:
                  properties:
                    recovery_codes:
                      items:
                        type: string
                      type: array
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Regenerate 2FA recovery codes
      tags:
      - 2fa
  /auth/2fa/setup:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.envelope'
            - properties:
                data:
                  properties:
                    otpauth_url:
                      type: string
                    qr_data_uri:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Begin 2FA setup — returns QR code
      tags:
      - 2fa
  /auth/2fa/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Challenge token and TOTP/recovery code
        in: body
        name: body
        required: true
        schema:
          properties:
            code:
              type: string
            mfa_challenge_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.envelope'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Complete 2FA login with TOTP code or recovery code
      tags:
      - 2fa
...
  /profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Get current user's profile
      tags:
      - profile
    patch:
      consumes:
      - application/json
      parameters:
      - description: Fields to update
        in: body
        name: body
        schema:
          properties:
            bio:
              type: string
            display_name:
              type: string
            first_name:
              type: string
            last_name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Update profile fields
      tags:
      - profile
  /profile/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Avatar image
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.envelope'
            - properties:
                data:
                  properties:
                    avatar_url:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Upload avatar image (multipart/form-data, max 2 MB)
      tags:
      - profile
  /profile/delete:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Schedule account for deletion (30-day window, sends cancellation email)
      tags:
      - profile
  /profile/delete/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Cancellation token
        in: body
        name: body
        required: true
        schema:
          properties:
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Cancel scheduled account deletion using token from email
      tags:
      - profile
  /profile/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: New email
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Request an email address change (sends verification to new address)
      tags:
      - profile
  /profile/email/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token and new email
        in: body
        name: body
        required: true
        schema:
          properties:
            new_email:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Confirm email change with token from verification email
      tags:
      - profile
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"